plotter/requirements.txt
plotly-express
dash
flask-caching
pyuwsgi
